Move CollectionOfIdentified Entries to own file
