+++ profile/src/v100/entity/persona/persona_data/mod.rs
@@ -1,3 +1,4 @@
+pub mod collection_of_identified_entries;
 pub mod entry;
 pub mod entry_kinds;
 pub mod identified_entry;
